Start refactoring
Cleaner code = healthier devs!
