+++ TFD_Units/Classes/CfgFactionClasses.hpp
@@ -1,0 +1,34 @@
+class CfgFactionClasses {
+
+	class TFD_AA {
+		displayName = "Armee de l'Air";
+		priority = 5;
+		side = 1;
+		icon = "\TFD_Units\Data\Icons\Ico_AA.paa";
+		//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+	};
+
+	class TFD_AT {
+		displayName = "Armee de Terre";
+		priority = 6;
+		side = 1;
+		icon = "\TFD_Units\Data\Icons\Ico_AT.paa";
+		//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+	};
+
+	class TFD_GN {
+		displayName = "Gendarmerie Nationale";
+		priority = 7;
+		side = 1;
+		icon = "\TFD_Units\Data\Icons\Ico_GN.paa";
+		//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+	};
+
+	class TFD_MN {
+		displayName = "Marine Nationale";
+		priority = 8;
+		side = 1;
+		icon = "\TFD_Units\Data\Icons\Ico_MN.paa";
+		//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+	};
+};
