+++ TFD_Units/Classes/CfgPatches.hpp
@@ -1,0 +1,10 @@
+class CfgPatches {
+	class TFD_Units {
+		units[] = {"TFD_SacIntervention","TFD_Gendarme_Depart","TFD_Gendarme_Mobile","TFD_Pompier_polo","TFD_PompierAir_polo","TFD_Vigile_polo","TFD_PolicierMun_poloMarine","TFD_PolicierMun_poloBlanc","TFD_PolicierMun_poloASVP","TFD_PolicierMun_poloRurale","TFD_SurveillantAdmPen_polo","TFD_Formateur_polo","TFD_Benevole_polo","TFD_Douanier_polo","TFD_SapeurSauveteur_polo","TFD_PolicierNat_Polo","TTFD_BenevoleCRF_polo","TFD_BenevolePC_polo","TFD_Pompier_PL","TFD_Pistard_AA","TFD_Patrouilleur_DIR","TFD_Marin_TPB","TFD_Marin_TPBfeu","TFD_Ambulancier_SAMU","TFD_Medecin_CICR","TFD_Chirurgien_CICR","TFD_Pilote_GN","TFD_Pilote_SC","TFD_Soldat_Pull","TFD_Aviateur_Pull","TFD_Plouf_UBACS"};
+		weapons[] = {"TFD_CasquetteGend_Depart","TFD_CasquetteGend_Mobile","TFD_CasquettePM_ASVP","TFD_CasquettePM","TFD_CasquetteAdmPen","TFD_CasquetteSP","TFD_CasquettePM_rurale","TFD_CasquetteSecurite","TFD_CasquetteCadet","TFD_CasquettePolice","TFD_CasquettePC","TFD_CasquetteABFormateur","TFD_CasqueF2_HDR","TFD_CasqueF2_SO","TFD_CasqueF2_Off","TFD_CasqueF2_PC","TFD_CasqueHeli_GN","TFD_CasqueHeli_SC","TFD_CasqueMoto_GN","TFD_RecycleurHV","TFD_GHV_Infirmier","TFD_GHV_Medecin","TFD_GHV_Formateur","TFD_GiletTac_GN","TFD_GiletTac_PN","TFD_GiletTac_ERIS","TFD_GiletTac_Nego","TFD_GiletTac_SC","TFD_GiletTac_CE","TFD_GiletS3_GN","TFD_GiletS3_PN","TFD_GiletS3_CE","TFD_GiletAssaut_GIPN","TFD_GiletAssaut_Douane","TFD_GiletAssaut_GN","TFD_GiletAssautLourd_GIPN","TFD_GiletAssautLourd_Douane","TFD_GiletAssautLourd_GN","TFD_PoloGend","TFD_PoloSP","TFD_PoloSP_air","TFD_PoloSecurite","TFD_PoloPM_marine","TFD_PoloPM_blanc","TFD_PoloPM_ASVP","TFD_PoloPM_rurale","TFD_PoloAdmPen","TFD_PoloFormateur","TFD_PoloBenevole","TFD_PoloDouane","TFD_PoloUIISC","TFD_PoloPN","TFD_PoloCRF","TFD_PoloPC","TFD_WetsuitHV","TFD_CombiPistard","TFD_CombiDIR","TFD_CombiTPB","TFD_CombiTPBfeu","TFD_CombiSAMU","TFD_CombiCICR","TFD_CombiCICR_sang","TFD_CombiPilote_GN","TFD_CombiPilote_SC","TFD_PullCE_AT","TFD_PullMarine_AA","TFD_UBACSCE_Palmes"};
+		requiredVersion = 0.1;
+		requiredAddons[] = {"A3_Characters_F_BLUFOR"};
+		version = 1.0;
+		author = "Heartbroken";
+	};
+};
